requirements.txt
attrs==19.3.0
boltons==20.0.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2020.4.5
chardet==3.0.4
click==7.1.1
coloredlogs==14.0
dependency-injector==3.15.6
entrypoints==0.3
firebase-admin==4.0.0
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-profiler==1.0.9
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.1
humanfriendly==8.1
idna==2.6
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.4.1
pytz==2019.3
requests==2.18.4
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.22
waitress==1.4.3
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
